add variable binding tests with condition tests
The test of the presence of a variable really has to do with the
structure of a condition, so add tests for it with the state/condition
tests. Make it clearer in the variables test file that it is testing
what characters can be in a variable, and add explanatory note to
one of the tests there.